drivers/drm/cdns-dsi-core: Temporarily disable warning
Whenever the kernel begins displaying anything on the DSI output, the first
attempt will dump a kernel warning. It's about the PLL not getting locked before
timeout, during stream enable of DSI. The warning only comes once, because it
uses WARN_ON_ONCE, but the PLL is not locked during every stream enable. This
was introduced when the DSI color shift patches were applied.  Not sure why
this happens, but it does not seem to affect the DSI display in any way. TI
will continue to debug this problem and it will get fixed in time.